add `ui/scrubber` tests
